requirements.txt
asgiref==3.8.1
colorama==0.4.6
confluent-kafka==2.3.0
Django==3.2.25
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
loguru==0.7.2
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
uritemplate==4.1.1
win32-setctime==1.1.0
psycopg2-binary==2.9.9